requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.1
pypdf>=4.2.0
langchain>=0.2.6
langchain-community>=0.2.6
langchain-mistralai>=0.1.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
numpy>=1.26.4
